requirements.txt
Django==3.2.15
psycopg2==2.9.3
djangorestframework==3.13.1
django-filter==22.1
environ==1.0
django-environ==0.9.0